source_archive/markusos-fork-snapshot/pyproject.toml
[project]
name = "kublet-apps"
version = "0.1.0"
description = "Tools and apps for the Kublet ESP32 display"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "esptool>=4.8.1",
    "esp-idf-nvs-partition-gen>=0.2.0",
    "pyserial",
]

[project.optional-dependencies]
dev = ["pytest", "ruff", "pillow"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools/src/kublet_dev"]

[tool.pytest.ini_options]
testpaths = ["tools/tests"]
pythonpath = ["tools/src"]

[tool.ruff]
target-version = "py313"
